Add files via upload
Each month, Interfaith Outreach analyzes the data on the number of hours each volunteer worked, the total number of volunteer hours each department completed, and other such data. Sometimes, volunteers forget to check-in and/or check-out. Manually searching for missing data is time-consuming, so I wrote an R script that automates this process and saves my supervisor time. Personal client information is removed or replaced with unique integers for this presentation.